+++ https-client/src/main/java/com/microej/example/iot/https/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright 2015-2023 MicroEJ Corp. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be found with this software.
+ */
+
+/**
+ * Contains the https page.
+ */
+@ej.annotation.NonNullByDefault
+package com.microej.example.iot.https;