+++ androidconnectivity/CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.0] - 2024-06-11
+
+### Changed
+
+- Update ``com.microej.gradle.application`` plugin version to 0.17.0.
+- Minor README updates.
+
 ## [4.0.0] - 2024-02-08
 
 ### Changed